+++ docs/changelog.rst
@@ -28,6 +28,8 @@ Improvements
 ~~~~~~~~~~~~
 - ```FeatureSelector``` is now compatible with Tree-structure Parzen Estimator
   method in Osprey (gh-1018).
+- Fixed bug in ```from_msm``` method for ```PCCA``` and ```PCCAPlus``` which
+  now allows a ```PCCAPlus``` objective function to be specified (gh-1036).
 - ```msmbuilder.io.sampling.sample_dimension``` with ```scheme='edge'``` now works properly. (#1043)